+++ benches/datatypes_benchmarks.rs
@@ -9,8 +9,6 @@ mod datatypes {
     pub mod cont;
     pub mod id;
     pub mod io;
-    pub mod iso_lens;
-    pub mod iso_prism;
     pub mod lazy_error;
     pub mod lens;
     pub mod prism;
@@ -29,8 +27,6 @@ use datatypes::composable_error::composable_error_benchmarks;
 use datatypes::cont::cont_benchmarks;
 use datatypes::id::id_benchmarks;
 use datatypes::io::io_benchmarks;
-use datatypes::iso_lens::iso_lens_benchmarks;
-use datatypes::iso_prism::iso_prism_benchmarks;
 use datatypes::lazy_error::lazy_error_benchmarks;
 use datatypes::lens::lens_benchmarks;
 use datatypes::prism::prism_benchmarks;
@@ -53,8 +49,6 @@ criterion_group!(
     io_benchmarks,
     lens_benchmarks,
     prism_benchmarks,
-    iso_lens_benchmarks,
-    iso_prism_benchmarks,
     pvec_benchmarks,
     composable_error_benchmarks,
     lazy_error_benchmarks,
@@ -73,8 +67,6 @@ criterion_group!(
     io_benchmarks,
     lens_benchmarks,
     prism_benchmarks,
-    iso_lens_benchmarks,
-    iso_prism_benchmarks,
     pvec_benchmarks,
     composable_error_benchmarks,
     lazy_error_benchmarks,